Two updates to the SAW integration:
1. Update the SAW download to use the latest release instead of an
ancient nightly that's no longer available.

2. Skip the downloads of both SAW and Cryptol if IN_SAW_CI is set
to "yes".

These are motivated by the use of this tree in the SAW CI tests, which
has been blowing up, first because we tried to delete the ancient
nightlies, then because the website reorg broke the downloads.

However, at least point (1) seems to be of general value.

+++ scripts/install.sh
@@ -8,6 +8,7 @@ ABC_URL='https://saw.galois.com/builds/abc/abc-c78ee311-Linux.tar.gz'
 
 if [ $# -ne 0 ] && [ "$1" = "--latest" ]; then
   # Determine the URL of the latest SAW and Cryptol nightly
+  # XXX: this will not work any more
   #########################################################
   SAW_DATE=$(curl -s https://saw.galois.com/builds/nightly/ | grep saw | grep -o "[0-9]\{4\}-[0-9]\{2\}-[0-9]\{2\}" | sort | tail -n 1) # `grep -o` says print only the matched substring
   SAW_NIGHTLY=$(curl -s https://saw.galois.com/builds/nightly/ | grep -oP "saw.*?${SAW_DATE}-Linux.*?\.tar\.gz"  | head -n 1) # `curl -s` means silent; `grep -o` says print only the matched substring; `grep -P` says Perl syntax, which we use to get a lazy match (shortest) with .*?
@@ -17,7 +18,8 @@ if [ $# -ne 0 ] && [ "$1" = "--latest" ]; then
   CRYPTOL_NIGHTLY=$(curl -s https://cryptol.net/builds/nightly/ | grep -oP "cryptol.*?${CRYPTOL_DATE}-Ubuntu.*?\.tar\.gz"  | head -n 1) # `curl -s` means silent; `grep -o` says print only the matched substring; `grep -P` says Perl syntax, which we use to get a lazy match (shortest) with .*?
   CRYPTOL_URL="https://cryptol.net/builds/nightly/${CRYPTOL_NIGHTLY}"
 else
-  SAW_URL="https://saw.galois.com/builds/nightly/saw-0.8.0.99-2021-12-06-Linux-x86_64.tar.gz"
+  #SAW_URL="https://saw.galois.com/builds/nightly/saw-0.8.0.99-2021-12-06-Linux-x86_64.tar.gz"
+  SAW_URL='https://github.com/GaloisInc/saw-script/releases/download/v1.2/saw-1.2-ubuntu-20.04-X64.tar.gz'
   CRYPTOL_URL="https://github.com/GaloisInc/cryptol/releases/download/2.11.0/cryptol-2.11.0-Linux-x86_64.tar.gz"
 
 fi
@@ -56,7 +58,7 @@ then
 fi
 
 # fetch SAW
-if [ ! -f bin/saw ]
+if [ ! -f bin/saw ] && [ ${IN_SAW_CI:-no} != yes ]
 then
     mkdir -p deps/saw
     wget $SAW_URL -O deps/saw.tar.gz
@@ -66,7 +68,7 @@ then
 fi
 
 # fetch Cryptol
-if [ ! -f bin/cryptol ]
+if [ ! -f bin/cryptol ] && [ ${IN_SAW_CI:-no} != yes ]
 then
     mkdir -p deps/cryptol
     wget $CRYPTOL_URL -O deps/cryptol.tar.gz
